+++ podcasts/Analytics/AnalyticsEvent.swift
@@ -695,4 +695,22 @@ enum AnalyticsEvent: String {
     case deselectChaptersChapterSelected
     case deselectChaptersChapterDeselected
 
+    // MARK: - Kids Profile
+    case kidsProfileBannerSeen
+    case kidsProfileEarlyAccessRequested
+    case kidsProfileBannerDismissed
+    case kidsProfileSendFeedbackTapped
+    case kidsProfileNoThankYouTapped
+    case kidsProfileThankYouForYourInterestSeen
+    case kidsProfileFeedbackFormSeen
+    case kidsProfileFeedbackSent
+
+    // MARK: - Transcript
+
+    case transcriptShown
+    case transcriptError
+    case transcriptDismissed
+    case transcriptSearch
+    case transcriptSearchNextResult
+    case transcriptSearchPreviousResult
 }