+++ podcasts/ProfileViewController.swift
@@ -287,7 +287,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         case .endOfYearPrompt:
             return EndOfYearPromptCell()
         case .bookmarks:
-            cell.settingsImage.image = UIImage(named: "bookmarks-shelf-icon")
+            cell.settingsImage.image = UIImage(named: "bookmarks-profile")
             cell.settingsLabel.text = L10n.bookmarks
         }
 
@@ -297,6 +297,13 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
     func tableView(_ tableView: UITableView, shouldHighlightRowAt indexPath: IndexPath) -> Bool {
         let row = tableData[indexPath.section][indexPath.row]
         return row != .kidsProfile
+    }
+
+    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
+        let row = tableData[indexPath.section][indexPath.row]
+        if row == .kidsProfile {
+            Analytics.track(.kidsProfileBannerSeen)
+        }
     }
 
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {