+++ podcasts/Sharing/ShareImageView.swift
@@ -24,6 +24,17 @@ enum ShareImageStyle: CaseIterable {
             return "small"
         }
     }
+
+    var videoSize: CGSize {
+        switch self {
+        case .large:
+            CGSize(width: 292, height: 438)
+        case .medium:
+            CGSize(width: 292, height: 293)
+        case .small:
+            CGSize(width: 324, height: 169)
+        }
+    }
 }
 
 struct ShareImageView: View {
@@ -31,39 +42,56 @@ struct ShareImageView: View {
     let info: ShareImageInfo
     let style: ShareImageStyle
 
+    @Binding var angle: Double
+
     var body: some View {
         ZStack {
             LinearGradient(gradient: info.gradient, startPoint: .top, endPoint: .bottom)
+            let rotationFactor = sin(Angle(degrees: angle).radians)
+            KidneyShape()
+                .fill(info.gradient.stops.first?.color ?? .black)
+                .blur(radius: 50)
+                .opacity(0.15 + 0.5 * abs(rotationFactor))
+                .offset(x: -50, y: 0)
+                .rotationEffect(.degrees(180 * rotationFactor))
+            KidneyShape()
+                .fill(.white)
+                .blur(radius: 50)
+                .offset(x: 50, y: 0)
+                .opacity(0.15 + 0.5 * abs(rotationFactor))
+                .rotationEffect(.degrees(-180 * rotationFactor))
+                .blendMode(.softLight)
             Color.black.opacity(0.2)
             switch style {
             case .large:
                 VStack(spacing: 32) {
                     image()
-                        .frame(width: 200, height: 200)
+                        .aspectRatio(1, contentMode: .fit)
                     text()
                     PocketCastsLogoPill()
                 }
                 .padding(24)
-                .frame(width: 292, height: 438)
+                .aspectRatio(0.66, contentMode: .fit)
             case .medium:
                 VStack(spacing: 24) {
                     image()
-                        .frame(width: 120, height: 120)
+                        .aspectRatio(1, contentMode: .fit)
                     text()
                         .frame(alignment: .leading)
                 }
                 .padding(24)
-                .frame(width: 292, height: 293)
+                .aspectRatio(0.99, contentMode: .fit)
             case .small:
                 HStack(spacing: 18) {
                     image()
-                        .frame(width: 120, height: 120)
+                        .aspectRatio(1, contentMode: .fit)
                     text(alignment: .leading, textAlignment: .leading, lineLimit: 3)
                 }
                 .padding(24)
-                .frame(width: 324, height: 169)
+                .aspectRatio(1.97, contentMode: .fit)
             }
         }
+        .frame(width: style.videoSize.width, height: style.videoSize.height)
         .fixedSize()
     }
 
@@ -92,7 +120,45 @@ struct ShareImageView: View {
     }
 }
 
-extension ShareImageView {
+struct KidneyShape: Shape {
+    func path(in rect: CGRect) -> Path {
+        var path = Path()
+        path.move(to: CGPoint(x: rect.minX, y: rect.midY))
+        path.addCurve(
+            to: CGPoint(x: rect.maxX, y: rect.midY),
+            control1: CGPoint(x: rect.minX + rect.width * 0.2, y: rect.minY),
+            control2: CGPoint(x: rect.maxX - rect.width * 0.2, y: rect.minY)
+        )
+        path.addCurve(
+            to: CGPoint(x: rect.minX, y: rect.midY),
+            control1: CGPoint(x: rect.maxX, y: rect.maxY),
+            control2: CGPoint(x: rect.minX, y: rect.maxY)
+        )
+
+        // Indentation
+        let indentWidth = rect.width * 0.2
+        let indentHeight = rect.height * 0.3
+        path.move(to: CGPoint(x: rect.midX - indentWidth, y: rect.minY + rect.height * 0.3))
+        path.addCurve(
+            to: CGPoint(x: rect.midX + indentWidth, y: rect.minY + rect.height * 0.3),
+            control1: CGPoint(x: rect.midX - indentWidth / 2, y: rect.minY + indentHeight),
+            control2: CGPoint(x: rect.midX + indentWidth / 2, y: rect.minY + indentHeight)
+        )
+
+        return path
+    }
+}
+
+@available(iOS 16.0, *)
+extension ShareImageView: Transferable {
+    static var transferRepresentation: some TransferRepresentation {
+        DataRepresentation<Self>(exportedContentType: .png) { view in
+            try await view.snapshot().pngData().throwOnNil()
+        }
+    }
+}
+
+extension View {
     func itemProvider() -> NSItemProvider {
         let itemProvider = NSItemProvider()
         if #available(iOS 16.0, *) {
@@ -116,13 +182,13 @@ extension ShareImageView {
 let previewInfo = ShareImageInfo(name: "This American Life", title: "Dylan Field, Figma Co-founder, Talks Design, Economy, and life after failed Adobe acquisitions", description: Date().formatted(), artwork: URL(string: "https://static.pocketcasts.com/discover/images/280/3782b780-0bc5-012e-fb02-00163e1b201c.jpg")!, gradient: Gradient(colors: [Color.red, Color(hex: "620603")]))
 
 #Preview("large") {
-    ShareImageView(info: previewInfo, style: .large)
+    ShareImageView(info: previewInfo, style: .large, angle: .constant(0))
 }
 
 #Preview("medium") {
-    ShareImageView(info: previewInfo, style: .medium)
+    ShareImageView(info: previewInfo, style: .medium, angle: .constant(0))
 }
 
 #Preview("small") {
-    ShareImageView(info: previewInfo, style: .small)
+    ShareImageView(info: previewInfo, style: .small, angle: .constant(0))
 }